+++ libs/media/src/lib/components/upload/widget/upload-widget.module.ts
@@ -11,7 +11,7 @@ import { MaxLengthModule } from '@blockframes/utils/pipes/max-length.pipe';
 import { TaskProgressModule } from '../../../pipes/task-progress.pipe';
 import { TaskSnapshotModule } from '../../../pipes/task-state.pipe';
 import { ImageReferenceModule } from '@blockframes/media/directives/image-reference/image-reference.module';
-import { FileTypeImagePipeModule } from '@blockframes/media/pipes/file-type-image.pipe';
+import { FileNameModule } from '@blockframes/utils/pipes/fileName.pipe';
 
 // Material
 import { MatExpansionModule } from '@angular/material/expansion';
@@ -29,7 +29,7 @@ import { MatDividerModule } from '@angular/material/divider';
     MaxLengthModule,
     TaskProgressModule,
     TaskSnapshotModule,
-    FileTypeImagePipeModule,
+    FileNameModule,
 
     // Blockframes
     ImageReferenceModule,
